Use FastNoise for everything noisy. Reduce allocations. Noise changes with time.
The time-based noise changes aren't very clean; ideally you could pass a particular z-value (to be used as time is now) and the noise would use it, but currently it just uses the time since the program started.
PerlinNoiseGenerator isn't used anymore; you can remove it if you want.
Most of the blending methods now blend with their destination rather than a new Pixmap, since that matches how they are used, and creating tons of Pixmaps is slow when rendering new frames as often as possible.